An attempt to make the code modular
